+++ docs/src/main/scala/me/shadaj/slinky/docs/DocsPage.scala
@@ -75,12 +75,15 @@ import js.Dynamic.literal
 
 object DocsTree {
   val tree: Map[String, List[(String, String)]] = Map(
-    "Quick Start" -> List(
+    "Core Concepts" -> List(
       "Installation" -> "/docs/installation/",
       "Hello World!" -> "/docs/hello-world/",
       "The Tag API" -> "/docs/the-tag-api/",
       "Writing Components" -> "/docs/writing-components/",
       "External Components" -> "/docs/external-components/"
+    ),
+    "Advanced Guides" -> List(
+      "Scala.js React Interop" -> "/docs/scalajs-react-interop/"
     )
   )
 }
@@ -153,10 +156,10 @@ import DocsTree._
       div(style := literal(
         display = "flex",
         flexDirection = "row"
-      ))(
+      ), className := "docs-page")(
         div(style := literal(
           width = "calc(100% - 300px)"
-        ))(
+        ), className := "docs-content")(
           div(style := literal(maxWidth = "1400px"))(
             state.document.map { t =>
               Remark().use(ReactRenderer, literal(
@@ -171,7 +174,7 @@ import DocsTree._
         div(style := literal(
           width = "300px",
           marginLeft = "20px"
-        ))(
+        ), className := "docs-sidebar")(
           div(
             style := literal(
               position = "fixed",
@@ -182,7 +185,8 @@ import DocsTree._
               paddingTop = "40px",
               paddingRight = "1000px",
               boxSizing = "border-box"
-            )
+            ),
+            className := "docs-sidebar-content"
           )(
             nav(style := literal(
               position = "relative",